requirements.txt
ttkthemes==3.2.2
flask==3.0.2
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
werkzeug==3.0.1 